+++ test/inspec/author-publish-dispatcher_spec.rb
@@ -38,6 +38,18 @@ describe ssl(port: 5432) do
   it { should be_enabled }
 end
 
+# TODO: re-add this test after device name can be retrieved from a configuration
+# describe etc_fstab.where { device_name == '/dev/xvdb' } do
+#   its('mount_point') { should cmp '/mnt/ebs1' }
+# end
+
+describe file('/opt/aem/author') do
+  its('type') { should eq :symlink }
+  its('link_path') { should eq '/mnt/ebs1/author' }
+  its('owner') { should eq 'aem-author' }
+  its('group') { should eq 'aem-author' }
+end
+
 if %w[amazon].include?(os[:name]) && !os[:release].start_with?('20\d\d')
 
   describe systemd_service('aem-publish') do
@@ -69,6 +81,18 @@ end
 describe ssl(port: 5433) do
   it { should be_enabled }
 end
+
+describe file('/opt/aem/publish') do
+  its('type') { should eq :symlink }
+  its('link_path') { should eq '/mnt/ebs2/publish' }
+  its('owner') { should eq 'aem-publish' }
+  its('group') { should eq 'aem-publish' }
+end
+
+# TODO: re-add this test after device name can be retrieved from a configuration
+# describe etc_fstab.where { device_name == '/dev/xvdc' } do
+#   its('mount_point') { should cmp '/mnt/ebs2' }
+# end
 
 if %w[amazon].include?(os[:name]) && !os[:release].start_with?('20\d\d')